+++ packages/rx_bloc_cli/lib/src/processors/common/generated_files_processor.dart
@@ -2,8 +2,10 @@ import 'package:rx_bloc_cli/src/processors/android/app_build_gradle_processor.da
 import 'package:rx_bloc_cli/src/processors/common/file_string_processor.dart';
 
 import 'package:rx_bloc_cli/src/processors/ios/flutter_xcconfig_file_processor.dart';
+import 'package:rx_bloc_cli/src/rx_bloc_cli_constants.dart';
 
 import '../../models/generator_arguments.dart';
+import '../../utils/flavor_generator.dart';
 import '../android/android_build_gradle_processor.dart';
 import '../android/android_manifest_processor.dart';
 import '../ios/ios_runner_project_processor.dart';
@@ -51,18 +53,12 @@ class GeneratedFilesProcessor {
   /// located within the ios/Flutter directory
   List<(String, StringProcessor)>
       _buildIOSFlutterXCConfigFileProcessorTargets() => [
-            'devDebug',
-            'devProfile',
-            'devRelease',
-            'prodDebug',
-            'prodProfile',
-            'prodRelease',
-            'sitDebug',
-            'sitProfile',
-            'sitRelease',
-            'uatDebug',
-            'uatProfile',
-            'uatRelease',
+            ...FlavorGenerator.getFlavors(args)
+                .map((flavor) => kIOSBuildModes.map((mode) => '$flavor$mode'))
+                .fold(<String>[], (prev, cur) {
+              prev.addAll(cur);
+              return prev;
+            }),
           ]
               .map(
                 (file) => (