+++ packages/rx_bloc_cli/lib/src/processors/ios/podfile_processor.dart
@@ -1,4 +1,6 @@
 import 'package:rx_bloc_cli/src/extensions/string_buffer_extensions.dart';
+import 'package:rx_bloc_cli/src/rx_bloc_cli_constants.dart';
+import 'package:rx_bloc_cli/src/utils/flavor_generator.dart';
 
 import '../common/abstract_processors.dart';
 
@@ -8,11 +10,18 @@ class PodfileProcessor extends StringProcessor {
   /// Podfile file processor constructor
   PodfileProcessor(super.args);
 
+  Map<String, String> get _buildModeMapping => {
+        'Debug': 'debug',
+        'Profile': 'release',
+        'Release': 'release',
+      };
+
   @override
   String execute() {
     if (input == null) return '';
     final buffer = StringBuffer(input!);
 
+    _updateGlobalIOSPlatform(buffer);
     _updateProjectRunnerModes(buffer);
 
     return buffer.toString();
@@ -20,25 +29,41 @@ class PodfileProcessor extends StringProcessor {
 
   /// region Private methods
 
+  void _updateGlobalIOSPlatform(StringBuffer buffer) {
+    final replacement =
+        buffer.toString().replaceAll('# platform :ios', 'platform :ios');
+    buffer
+      ..clear()
+      ..write(replacement);
+  }
+
   void _updateProjectRunnerModes(StringBuffer buffer) {
+    const runnerConfigs = "project 'Runner'";
+    const runnerConfigsComment =
+        '# Map each flavor configuration to a build mode\n';
     final index = buffer.nthIndexOf('}',
             start: buffer.nthIndexOf(
               '{',
-              start: buffer.nthIndexOf("project 'Runner'"),
+              start: buffer.nthIndexOf(runnerConfigs),
             )) -
         1;
-    buffer.insertAt(index, _buildConfigForSelectedFlavors());
+    buffer
+      ..insertAt(index, _buildConfigForSelectedFlavors())
+      ..insertBefore(runnerConfigs, runnerConfigsComment);
   }
 
-  String _buildConfigForSelectedFlavors() => {'prod', 'dev', 'sit', 'uat'}.fold(
-      '',
-      (previousValue, element) =>
-          previousValue + _buildConfigForFlavor(element));
+  String _buildConfigForSelectedFlavors() =>
+      FlavorGenerator.getFlavors(args).fold(
+          '',
+          (previousValue, element) =>
+              previousValue + _buildConfigForFlavor(element));
 
-  String _buildConfigForFlavor(String flavor) =>
-      '''\n  'Debug-$flavor' => :debug,
-  'Profile-$flavor' => :release,
-  'Release-$flavor' => :release,''';
+  String _buildConfigForFlavor(String flavor) => kIOSBuildModes
+      .map((mode) => _buildSingleConfigForFlavor(flavor, mode))
+      .fold('', (prev, cur) => prev + cur);
+
+  String _buildSingleConfigForFlavor(String flavor, String buildMode) =>
+      "\n  '$buildMode-$flavor' => :${_buildModeMapping[buildMode]},";
 
   /// endregion
 }