+++ packages/rx_bloc_cli/mason_templates/bricks/rx_bloc_flavor_config/__brick__/setup_scripts/firebase_build_script.sh
@@ -1,22 +1,22 @@
 # Copies appropriate Firebase configuration files based on the current run configuration
 
-if [ "${CONFIGURATION}" == "Debug-prod" ] || [ "${CONFIGURATION}" == "Release-prod" ] || [ "${CONFIGURATION}" == "Release" ]; then
+if [ "${CONFIGURATION}" == "Debug-production" ] || [ "${CONFIGURATION}" == "Profile-production" ] || [ "${CONFIGURATION}" == "Release-production" ] || [ "${CONFIGURATION}" == "Profile" ] || [ "${CONFIGURATION}" == "Release" ]; then
 cp -r "${PROJECT_DIR}/environments/production/firebase/GoogleService-Info.plist" "${PROJECT_DIR}/Runner/GoogleService-Info.plist"
 
 echo "Production plist copied"
-elif [ "${CONFIGURATION}" == "Debug-sit" ] || [ "${CONFIGURATION}" == "Release-sit" ]; then
+elif [ "${CONFIGURATION}" == "Debug-sit" ] || [ "${CONFIGURATION}" == "Profile-sit" ] || [ "${CONFIGURATION}" == "Release-sit" ]; then
 
 cp -r "${PROJECT_DIR}/environments/sit/firebase/GoogleService-Info.plist" "${PROJECT_DIR}/Runner/GoogleService-Info.plist"
 
 echo "SIT plist copied"
 
-elif [ "${CONFIGURATION}" == "Debug-uat" ] || [ "${CONFIGURATION}" == "Release-uat" ]; then
+elif [ "${CONFIGURATION}" == "Debug-uat" ] || [ "${CONFIGURATION}" == "Profile-uat" ] || [ "${CONFIGURATION}" == "Release-uat" ]; then
 
 cp -r "${PROJECT_DIR}/environments/uat/firebase/GoogleService-Info.plist" "${PROJECT_DIR}/Runner/GoogleService-Info.plist"
 
 echo "UAT plist copied"
 
-elif [ "${CONFIGURATION}" == "Debug-dev" ] || [ "${CONFIGURATION}" == "Release-dev" ] || [ "${CONFIGURATION}" == "Debug" ]; then
+elif [ "${CONFIGURATION}" == "Debug-development" ] || [ "${CONFIGURATION}" == "Profile-development" ] || [ "${CONFIGURATION}" == "Release-development" ] || [ "${CONFIGURATION}" == "Debug" ]; then
 
 cp -r "${PROJECT_DIR}/environments/development/firebase/GoogleService-Info.plist" "${PROJECT_DIR}/Runner/GoogleService-Info.plist"