+++ packages/rx_bloc_cli/mason_templates/bricks/lib_auth_matrix/__brick__/lib/lib_auth_matrix/views/auth_matrix_pin_biometrics_page.dart
@@ -21,14 +21,12 @@ class AuthMatrixPinBiometricsPage extends StatelessWidget {
   final String endToEndId;
 
   @override
-  Widget build(BuildContext context) => WillPopScope(
-        onWillPop: () async {
-          context
-              .read<PinOtpBlocType>()
-              .events
-              .cancelAuthMatrix(response, endToEndId);
-          return true;
-        },
+  Widget build(BuildContext context) => PopScope(
+        canPop: true,
+        onPopInvoked: (didPop) => context
+            .read<PinOtpBlocType>()
+            .events
+            .cancelAuthMatrix(response, endToEndId),
         child: Scaffold(
           appBar: AppBar(
             title: Text(context.l10n.featureAuthMatrix.pinBiometrics),