+++ packages/rx_bloc_cli/mason_templates/bricks/lib_pin_code/__brick__/lib/lib_pin_code/views/update_pin_page.dart
@@ -43,11 +43,12 @@ class _UpdatePinPageState extends State<UpdatePinPage> {
 
   @override
   Widget build(BuildContext context) => Builder(
-        builder: (context) => WillPopScope(
-          onWillPop: () async {
-            context.read<UpdateAndVerifyPinBlocType>().events.deleteSavedData();
-            return true;
-          },
+        builder: (context) => PopScope(
+          canPop: true,
+          onPopInvoked: (didPop) => context
+              .read<UpdateAndVerifyPinBlocType>()
+              .events
+              .deleteSavedData(),
           child: Scaffold(
             appBar: AppBar(
               title: Text(