+++ packages/rx_bloc_cli/mason_templates/bricks/lib_pin_code/__brick__/lib/lib_pin_code/views/verify_pin_code_page.dart
@@ -42,11 +42,12 @@ class _VerifyPinCodePageState extends State<VerifyPinCodePage> {
 
   @override
   Widget build(BuildContext context) => Builder(
-        builder: (context) => WillPopScope(
-          onWillPop: () async {
-            context.read<UpdateAndVerifyPinBlocType>().events.deleteSavedData();
-            return true;
-          },
+        builder: (context) => PopScope(
+        canPop: true,
+        onPopInvoked: (didPop) => context
+            .read<UpdateAndVerifyPinBlocType>()
+            .events
+            .deleteSavedData(),
           child: Scaffold(
             appBar: AppBar(
               title: Text(